Updating the excel file handling to save records into the history table. Refactoring.
